add 'delete category'

+++ security/src/main/java/vn/iotstar/security/service/impl/CategoryServiceImpl.java
@@ -11,7 +11,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.ObjectUtils;
 
 import vn.iotstar.security.model.Category;
+import vn.iotstar.security.model.Product;
 import vn.iotstar.security.repository.CategoryRepository;
+import vn.iotstar.security.repository.ProductRepository;
 import vn.iotstar.security.service.CategoryService;
 
 
@@ -20,6 +22,9 @@ public class CategoryServiceImpl implements CategoryService {
 
 	@Autowired
 	private CategoryRepository categoryRepository;
+	
+	@Autowired
+	private ProductServiceImpl productService;
 
 	@Override
 	public Category saveCategory(Category category) {
@@ -39,6 +44,10 @@ public class CategoryServiceImpl implements CategoryService {
 	@Override
 	public Boolean deleteCategory(int id) {
 		Category category = categoryRepository.findById(id).orElse(null);
+		
+		if(!productService.deleteProductByCategory(category)) {
+			return false;
+		}		
 
 		if (!ObjectUtils.isEmpty(category)) {
 			categoryRepository.delete(category);
